Fix: unfinished string near '"' when packing
This commit have a lot of changes in both (packer and linker)

The code was refactored, making it easier to read and maintain.

In addition, both the linker and the packer were updated to prevent
bad code generation when creating embedded scripts.

When the code was packaged with the packer, the mark was generated
together with important opcodes in the binary file, corrupting it
and preventing its full functioning.

During the linking stage, sanitization was added to the shebangs,
automatically removing them.

A summary of the changes:

- mark added automatically by the packer
- sanitization of shebangs by the linker
- linked and packaged files are now generated with namespace to
  avoid accidental overwriting